学习demo

+++ src/main/java/com/xie/reflect/GoodStudent.java
@@ -9,5 +9,5 @@ public interface GoodStudent {
 
     void studyHard(String book);
 
-    void studyHard(String book,int score);
+    void studyHard(String book,float score);
 }
